pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-risk-sentinel"
version = "0.1.0"
description = "Multi-Agent System for Proactive AI Risk Documentation Analysis and Correction"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GenAISafety Team", email = "team@genaisafety.com"},
    {name = "Preventera", email = "contact@preventera.com"}
]
keywords = ["ai", "risk", "safety", "hse", "osh", "ml", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.2.4",
    
    # Task Queue
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # ML & NLP
    "sentence-transformers>=2.3.0",
    "torch>=2.1.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    
    # API Clients
    "huggingface-hub>=0.20.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    
    # LLM Integration
    "anthropic>=0.18.0",
    "openai>=1.10.0",
    
    # Data Processing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "python-multipart>=0.0.6",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "rapidfuzz>=3.6.0",
    
    # Monitoring
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
ai-risk-sentinel = "ai_risk_sentinel.cli:app"
ars = "ai_risk_sentinel.cli:app"

[project.urls]
Homepage = "https://github.com/Preventera/AI-Risk-Sentinel"
Documentation = "https://preventera.github.io/AI-Risk-Sentinel"
Repository = "https://github.com/Preventera/AI-Risk-Sentinel"
Issues = "https://github.com/Preventera/AI-Risk-Sentinel/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_risk_sentinel"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/data",
]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hatch
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["ai_risk_sentinel"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "ARG001", # unused function argument
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG001", "S101"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src/ai_risk_sentinel",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/ai_risk_sentinel"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
